refactor: organize affiliates into a package
