+++ scripts/affiliates.py
@@ -1,5 +1,6 @@
-from yearn.affiliates import process_affiliates
+from yearn.affiliates.partners import partners
+from yearn.affiliates.snapshot import process_affiliates
 
 
 def main():
-    process_affiliates()
+    process_affiliates(partners)
