+++ yearn/affiliates/partners.py
@@ -1,0 +1,80 @@
+from yearn.affiliates.snapshot import Partner, Wrapper
+
+
+partners = [
+    Partner(
+        name='alchemix',
+        treasury='0x8392F6669292fA56123F71949B52d883aE57e225',
+        wrappers=[
+            Wrapper(
+                name='dai',
+                vault='0x19D3364A399d251E894aC732651be8B0E4e85001',
+                wrapper='0x014dE182c147f8663589d77eAdB109Bf86958f13',
+            ),
+            Wrapper(
+                name='dai-2',
+                vault='0x19D3364A399d251E894aC732651be8B0E4e85001',
+                wrapper='0x491EAFC47D019B44e13Ef7cC649bbA51E15C61d7',
+            ),
+        ],
+    ),
+    Partner(
+        name='inverse',
+        treasury='0x926dF14a23BE491164dCF93f4c468A50ef659D5B',
+        wrappers=[
+            Wrapper(
+                name='dai-wbtc',
+                vault='0x19D3364A399d251E894aC732651be8B0E4e85001',
+                wrapper='0xB0B02c75Fc1D07d351C991EBf8B5F5b48F24F40B',
+            ),
+            Wrapper(
+                name='dai-weth',
+                vault='0x19D3364A399d251E894aC732651be8B0E4e85001',
+                wrapper='0x57faa0dec960ed774674a45d61ecfe738eb32052',
+            ),
+            Wrapper(
+                name='usdc-weth',
+                vault='0x5f18C75AbDAe578b483E5F43f12a39cF75b973a9',
+                wrapper='0x698c1d40574cd90f1920f61D347acCE60D3910af',
+            ),
+        ],
+    ),
+    Partner(
+        name='frax',
+        treasury='0x8d0C5D009b128315715388844196B85b41D9Ea30',
+        wrappers=[
+            Wrapper(
+                name='usdc',
+                vault='0x5f18C75AbDAe578b483E5F43f12a39cF75b973a9',
+                wrapper='0xEE5825d5185a1D512706f9068E69146A54B6e076',
+            ),
+        ],
+    ),
+    Partner(
+        name='pickle',
+        treasury='0x066419EaEf5DE53cc5da0d8702b990c5bc7D1AB3',
+        wrappers=[
+            Wrapper(
+                name='usdc',
+                vault='0x5f18C75AbDAe578b483E5F43f12a39cF75b973a9',
+                wrapper='0xEecEE2637c7328300846622c802B2a29e65f3919',
+            ),
+            Wrapper(
+                name='lusd',
+                vault='0x5fA5B62c8AF877CB37031e0a3B2f34A78e3C56A6',
+                wrapper='0x699cF8fE0C1A6948527cD4737454824c6E3828f1',
+            ),
+        ],
+    ),
+    Partner(
+        name='badger',
+        treasury='0xB65cef03b9B89f99517643226d76e286ee999e77',
+        wrappers=[
+            Wrapper(
+                name='wbtc',
+                vault='0xA696a63cc78DfFa1a63E9E50587C197387FF6C7E',
+                wrapper='0x4b92d19c11435614CD49Af1b589001b7c08cD4D5',
+            ),
+        ],
+    ),
+]
